requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
circuitbreaker==1.4.0
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
oci==2.110.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
zipp==3.17.0
